fixed checkout error
