+++ src/js/models/gateway.js
@@ -22,7 +22,7 @@ var GatewayModel = PageModel.extend({
         
 
         // track order with google analytics
-//                this.trackOrder(order);
+//		this.trackOrder(order);
 
         // navigate to confirmation page
         einzl.router.navigate('/confirmation/' + order.id, {
@@ -86,6 +86,36 @@ var GatewayModel = PageModel.extend({
             }
             
         });
+		
+	},
+    
+    trackOrder: function(order) {
+		
+		ga('require', 'ecommerce');
+	
+		ga('ecommerce:addTransaction', {
+			'id': order.id,                     	// Transaction ID. Required.
+			'affiliation': 'Einzelstück',   		// Affiliation or store name.
+			'revenue': order.total,               	// Grand Total.
+			'shipping': order.shipping_price,       // Shipping.
+			'tax': '0',                     		// Tax.
+			'currency': 'EUR'						// Currency
+		});
+
+		$.each(order.cart.contents, function(key, item) {
+
+			ga('ecommerce:addItem', {
+				'id': key,                     		// Transaction ID. Required.
+				'name': item.name,    				// Product name. Required.
+				'sku': item.sku,                 	// SKU/code.
+				'category': item.category.value,    // Category or variation.
+				'price': item.total,                // Unit price.
+				'quantity': item.quantity           // Quantity.
+			});
+
+		});
+
+		ga('ecommerce:send');
 		
 	}
